requirements.txt
Flask==3.0.0
firebase-admin==6.4.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
Flask-Compress==1.24
